requirements.txt
dash-design-kit==1.6.7
dash==2.3.1
gunicorn==20.0.4
pandas==1.1.4
redis
google-cloud-bigquery
db-dtypes
colorcet
cryptography
